+++ CHANGELOG.md
@@ -6,6 +6,9 @@ It is based on Tidepool Blip 1.27.
 ### Added
 - YLP-2484 Remove level column in device change parameter history
 
+### Fixed
+- YLP-2572 Prevent patient from adding the same caregiver to their list twice
+
 ## 3.0.4-beta - 2023-02-10
 ### Added
 - YLP-1503 Logout idle user after 30 mn of inactivity